Fix stories file structure
